requirements.txt
pandas
numpy
python-dotenv
psycopg2-binary
sqlalchemy
scikit-learn
joblib
matplotlib
seaborn
openpyxl
